+++ educa100/src/main/java/com/educa100/controller/dto/response/MateriaResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record MateriaResponse(int id_materia) {
+public record MateriaResponse(Long id_materia) {
 }
