+++ educa100/src/main/java/com/educa100/controller/dto/response/UsuarioResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record UsuarioResponse(int id_usuario) {
+public record UsuarioResponse(Long id_usuario) {
 }
