+++ educa100/src/main/java/com/educa100/datasource/entity/MateriaEntity.java
@@ -9,12 +9,12 @@ import lombok.Data;
 public class MateriaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(unique = true)
     private String nome;
 
     @ManyToOne
     @JoinColumn(name = "id_curso", nullable = false)
-    private int id_curso;
+    private Long id_curso;
 }
