+++ educa100/src/main/java/com/educa100/datasource/repository/UsuarioRepository.java
@@ -15,6 +15,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
     Optional<UsuarioEntity> findByLogin(String login);
 }
