+++ educa100/src/main/java/com/educa100/infra/exception/AlunoNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class AlunoNotFoundException extends NotFoundException{
-    public AlunoNotFoundException(int id){
+    public AlunoNotFoundException(Long id){
         super("Aluno com o id " +id +" não encontrado");
     }
 }
