+++ educa100/src/main/java/com/educa100/infra/exception/CursoNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class CursoNotFoundException extends NotFoundException{
-    public CursoNotFoundException(int id){
+    public CursoNotFoundException(Long id){
         super("Curso com o id " +id +" não encontrado");
     }
 }
