+++ educa100/src/main/java/com/educa100/infra/exception/UsuarioNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class UsuarioNotFoundException extends NotFoundException{
-    public UsuarioNotFoundException(int id){
+    public UsuarioNotFoundException(Long id){
         super("Usuario com o id " +id +" não encontrado");
     }
 }
